python/consumer/pyproject.toml
[tool.poetry]
name = "consumer"
version = "0.1.0"
description = ""
authors = ["Amit Saha <amitsaha.in@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
kafka-python = "^1.4.7"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
